requirements.txt
docx2pdf
python-docx
openpyxl
python-pptx
Pillow
reportlab
xlsxwriter
pypandoc